+++ server/src/initGdc.termdb.js
@@ -1,61 +1,17 @@
 import ky from 'ky'
-import path from 'path'
+import { joinUrl } from './helpers'
 import { isUsableTerm } from '#shared/termdb.usecase.js'
 import serverconfig from './serverconfig.js'
 import { cachedFetch } from './utils'
 import { deepEqual } from '#shared/helpers.js'
 
 /*
-********************   Comment   *****************
-this script only runs once, when PP server/container launches.
-it is conditionally called by mds3.init.js when gdc dataset is present
-it caches publicly available info from some GDC apis, which doesn't require user token
-it is not a route, it does not accept request parameters from client
-upon any error, it throws exception and aborts launch.
-any error is considered critical and must be presented in server log for diagnosis
-
-
 ********************   functions    *************
-initGDCdictionary
+buildGDCdictionary
+	mayAddTermAttribute
 	assignDefaultBins
 	makeTermdbQueries
-	getOpenProjects
-	runRemainingWithoutAwait
-		testGDCapi
-			testRestApi
-			testGraphqlApi
-		mayCacheSampleIdMappingWithRegularCheck
-			cacheSampleIdMapping
-				fetchIdsFromGdcApi
-			mayReCacheCaseIdMapping
-				caseCacheIsStale
 
-
-******************** major tasks *****************
-- parsing gdc variables and constructing in-memory termdb:
-  HARDCODED LOGIC, does not need any configuration in dataset file
-  standard termdb "interface" functions are added to ds.cohort.termdb.q{}
-
-- determine default binconfig for all numeric terms
-  added to term json objects
-
-- querying list of open-access projects
-  stores at: ds.gdcOpenProjects = set of project ids that are open-access
-
-- test gdc api, make sure they're all online
-
-- cache sample/case name/uuid mapping
-  creates these new dataset-level attributes
-  ds.__gdc {
-  	aliquot2submitter{ get() }
-  	map2caseid{ get() }
-  	doneCaching: boolean, falg to indicate when the sample ID caching is done
-	casesWithExpData Set
-	caseid2submitter
-	data_release_version
-  }
-
-- periodic check of stale cache and re-cache above
 
 
 sections from api return that are used to build in-memory termdb
@@ -154,10 +110,8 @@ re.expand: []
 
 // prefix for keys in re._mapping{}
 const mapping_prefix = 'ssm_occurrence_centrics'
-// wait time for next check on stale case-id cache, 5min. feature flag allows testing with short internal
-const cacheCheckWait = serverconfig.features.gdcCacheCheckWait || 5 * 60 * 1000
 
-export async function initGDCdictionary(ds) {
+export async function buildGDCdictionary(ds) {
 	const id2term = new Map()
 	// k: term id, string with full path
 	// v: term obj
@@ -183,7 +137,7 @@ export async function initGDCdictionary(ds) {
 
 	// TODO switch to https://api.gdc.cancer.gov/cases/_mapping
 	const { host, headers } = ds.getHostHeaders()
-	const dictUrl = path.join(host.rest, 'ssm_occurrences/_mapping')
+	const dictUrl = joinUrl(host.rest, 'ssm_occurrences/_mapping')
 
 	let re
 	try {
@@ -358,17 +312,6 @@ export async function initGDCdictionary(ds) {
 	if (survivalCount) {
 		ds.cohort.termdb.termtypeByCohort.push({ cohort: '', termType: 'survival', termCount: survivalCount })
 		ds.cohort.termdb.termtypeByCohort.nested[''].survival = survivalCount
-	}
-
-	if (serverconfig.features.await4completeGdcCaseCache) {
-		// only use on dev environment. here as soon as server is launched,
-		// it will signal client to refresh (sse). if still caching asyncly,
-		// a gdc view may break due to incomplete cache, thus await a bit for cache to complete.
-		// also should use extApiCache
-		await runRemainingWithoutAwait(ds)
-	} else {
-		// use on prod, not to hold up container launch by caching
-		runRemainingWithoutAwait(ds)
 	}
 }
 
@@ -765,475 +708,5 @@ function makeTermdbQueries(ds, id2term) {
 				}
 			}
 		}
-	}
-}
-
-/****************************************************
-   remaining tasks after dictionary is built:
-   1. cache gdc case id mapping, cases with exp data
-   2. open-access projects
-   3. api testing
-   4. continuous check for stale cache
-*****************************************************
-
-Important!
-do not await when calling this function, as following steps execute logic that are optional for server function, and should not halt server launch
-should await each step so they work in sequence
-*/
-async function runRemainingWithoutAwait(ds) {
-	try {
-		await testGDCapi(ds)
-	} catch (e) {
-		// on some pp env e.g. ppirt, this api test sometimes fails. in order not to abort pp launch, only logs err out
-		// no need to record this test failure and report via healthcheck: since this test runs just once on launch. during the container lifetime it never runs again and thus is not informative of actual api status
-		console.log(e.message || e)
-	}
-
-	// may do it because it could be disabled by feature toggle
-	// caching action is fine-tuned by the feature toggle on a pp instance; log out detailed status per setting
-	if (serverconfig.features.stopGdcCacheAliquot) {
-		// do not cache at all. this flag is auto-set for container validation. running stale cache check will cause the server process not to quit, and break validation, thus must skip this when flag is true
-		console.log('GDC: sample IDs are not cached! No periodic check will take place!')
-		getCacheRef(ds) // though nothing is cached, must init the cache holder so not to break code that accesses this holder
-		///////////////////// NOTE ///////////////////////
-		// with missing cache for case id mapping and cases with exp data, any query using gdc gene exp data will not work!!
-		// this should be fine for container validation, but may not do so on a dev environment
-		return
-	}
-
-	try {
-		// obtain case id mapping for the first time and store at ds.__gdc
-		await cacheMappingOnNewRelease(ds)
-		// use setInterval instead of using setTimeout within cacheMappingOnNewRelease(),
-		// which will require separate setTimeout() calls at the end of the function
-		// plus within each error catch block
-		setInterval(cacheMappingOnNewRelease, cacheCheckWait, ds)
-	} catch (e) {
-		if (e.stack) console.log(e.stack)
-		throw 'cacheSampleIdMapping() failed: ' + (e.message || e)
-	}
-	// add any other gdc stuff
-}
-
-async function getOpenProjects(ds, ref) {
-	const data = {
-		filters: {
-			op: 'and',
-			content: [
-				{
-					op: '=',
-					content: {
-						field: 'access',
-						value: 'open'
-					}
-				},
-				{
-					op: '=',
-					content: {
-						field: 'data_type',
-						value: 'Masked Somatic Mutation'
-					}
-				}
-			]
-		},
-		facets: 'cases.project.project_id',
-		size: 0
-	}
-
-	const { host, headers } = ds.getHostHeaders()
-	const { body: re } = await cachedFetch(path.join(host.rest, 'files'), { method: 'POST', headers, body: data })
-
-	if (!Array.isArray(re?.data?.aggregations?.['cases.project.project_id']?.buckets)) {
-		console.log("getting open project_id but return is not re.data.aggregations['cases.project.project_id'].buckets[]")
-		return
-	}
-	for (const b of re.data.aggregations['cases.project.project_id'].buckets) {
-		// key is project_id value
-		if (b.key) ref.gdcOpenProjects.add(b.key)
-	}
-	console.log('GDC open-access projects:', ref.gdcOpenProjects.size)
-}
-
-/* this function is called when the gdc mds3 dataset is initiated on a pp instance
-primary purpose is to catch malformed api URLs on dev environment
-when running this on sj prod server, the gdc api can be down due to maintainance, and we do not want to prevent our server from launching
-thus do not halt process if api is down
-*/
-async function testGDCapi(ds) {
-	const { host, headers } = ds.getHostHeaders()
-	try {
-		// all these apis are available on gdc production and are publicly available
-		// if any returns an error code, will abort
-		{
-			const u = path.join(host.rest, 'ssms')
-			const c = await testRestApi(u)
-			if (c) throw `${u} returns error code: ${c}`
-		}
-		{
-			const u = path.join(host.rest, 'ssm_occurrences')
-			const c = await testRestApi(u)
-			if (c) throw `${u} returns error code: ${c}`
-		}
-		{
-			const u = path.join(host.rest, 'cases')
-			const c = await testRestApi(u)
-			if (c) throw `${u} returns error code: ${c}`
-		}
-		{
-			const u = path.join(host.rest, 'files')
-			const c = await testRestApi(u)
-			if (c) throw `${u} returns error code: ${c}`
-		}
-		{
-			const u = path.join(host.rest, 'analysis/top_mutated_genes_by_project')
-			const c = await testRestApi(u)
-			if (c) throw `${u} returns error code: ${c}`
-		}
-
-		// /data/ and /slicing/view/ are not tested as they require file uuid which is unstable across data releases
-
-		{
-			const c = await testGraphqlApi(host.graphql, JSON.parse(JSON.stringify(headers)))
-			if (c) throw `${host.graphql} returns error code: ${c}`
-		}
-	} catch (e) {
-		console.log(e)
-		throw `
-##########################################
-#
-#   Some GDC API unavailable, see error above
-#   ${host.rest}
-#   ${host.graphql}
-#
-##########################################`
-	}
-}
-
-/*
-if url is accessible, do not return
-if gets error:
-	return error code if available, so downstream logic can further act on the code
-	if no error code, throw and abort
-*/
-async function testRestApi(url) {
-	try {
-		const t = new Date()
-		const response = await ky(url)
-		if (response.status > 399) {
-			// 400 and above are all error, do not use !=200 as redirect (300+) is allowed...
-			// TODO console log additional response msg to help diagnosis
-			return response.status
-		}
-		console.log('GDC API okay: ' + url, new Date() - t, 'ms')
-	} catch (e) {
-		// if gdc service is down
-		console.log('See error from', url)
-		console.log(e)
-		if (e.code) return e.code
-		throw 'gdc api down: ' + url // no code, just throw to abort
-	}
-}
-
-async function testGraphqlApi(url, headers) {
-	// FIXME lightweight method to test if graphql is accessible?
-	const t = new Date()
-	try {
-		const query = `query termislst2total( $filters: FiltersArgument) {
-		explore {
-			cases {
-				aggregations (filters: $filters, aggregations_filter_themselves: true) {
-					primary_site {buckets { doc_count, key }}
-				}
-			}
-		}}`
-		const response = await ky.post(url, {
-			timeout: false,
-			headers,
-			json: { query, variables: {} }
-		})
-		if (response.status > 399) {
-			// 400 and above are all error, do not use !=200 as redirect (300+) is allowed...
-			// TODO console log additional response msg to help diagnosis
-			return response.status
-		}
-	} catch (e) {
-		console.log(e)
-		if (e.code) return e.code
-		throw 'see above error from graphql API ' + url
-	}
-	console.log('GDC GraphQL API okay: ' + url, new Date() - t, 'ms')
-}
-
-function getCacheRef(ds) {
-	// gather these arbitrary gdc stuff under __gdc{} to be safe
-	// do not freeze the object; they will be rewritten if cache is stale
-	const ref = {
-		aliquot2submitter: {
-			cache: new Map(),
-			get: async aliquot_id => {
-				if (ref.aliquot2submitter.cache.has(aliquot_id)) return ref.aliquot2submitter.cache.get(aliquot_id)
-
-				/* 
-				as on the fly api query is still slow, especially to query one at a time for hundreds of ids
-				simply return unconverted id to preserve performance
-				*/
-				return aliquot_id
-
-				// converts one id on the fly while the cache is still loading
-				//await fetchIdsFromGdcApi(ds, null, null, ref, aliquot_id)
-				//return ref.aliquot2submitter.cache.get(aliquot_id)
-			}
-		},
-		// create new attr to map to case uuid, from 3 kinds of ids: aliquot, sample submitter, and case submitter
-		// this mapping serves case selection from mds3 and matrix, where input can be one of these different types
-		map2caseid: {
-			cache: new Map(),
-			get: input => {
-				return ref.map2caseid.cache.get(input)
-				// NOTE if mapping is not found, do not return input, caller will call convert2caseId() to map on demand
-			}
-		},
-		caseid2submitter: new Map(), // k: case uuid, v: case submitter id
-		caseIds: new Set(), //
-		casesWithExpData: new Set(),
-		gdcOpenProjects: new Set(), // names of open-access projects
-		doneCaching: false
-	}
-
-	return ref
-}
-
-/*
-cache gdc sample id mappings
-** this is an optional step and can be skipped on dev machines **
-- create a map from sample aliquot id to sample submitter id, for displaying in mds3 tk
-- create a map from differet ids to case uuid, for creating gdc cohort with selected samples
-- cache list of case uuids with expression data
-
-function will rerun when it detects stale case id cache
-*/
-async function cacheMappingOnNewRelease(ds) {
-	console.log('GDC: checking if cache is stale')
-	// to avoid issues from race condition:
-	// - do not set ds.__gdc until the caching is complete
-	// - create a new ref object to map pending cacheable data
-	// - if a new cacheMapping call is triggered before the previous one is finished,
-	//   then make sure the previous cacheRef object is not used as ds.__gdc
-	const ref = getCacheRef(ds) // a new empty nested cache object
-	if (!ds.__gdc) ds.__gdc = ref // would reference the same object only on initial call
-
-	let version
-	try {
-		const response = await ds.preInit.getStatus()
-		version = response.data_release_version
-		if (deepEqual(version, ds.__pendingCacheVersion) || deepEqual(version, ds.__gdc.data_release_version)) {
-			if (ds.preInit.test)
-				console.log(
-					'GDC: skip recache of ',
-					version.minor,
-					ds.__pendingCacheVersion?.minor,
-					ds.__gdc.data_release_version.minor
-				)
-			// do not trigger duplicate caching for the same release version, whether pending or completed
-			return
-		}
-		// need to check before resetting ds.__pendingCacheVersion in subsequent lines
-		if (mayCancelStalePendingCache(ds, { data_release_version: version })) return
-		// not using deepEqual() here, since on initial call ds.__gdc and ref directly reference the same object
-		if (ds.__gdc !== ref) console.log('GDC: cache is stale. Re-caching...', version.minor)
-
-		// reset doneCaching to false, which may overwrite it for a previous data version that completed caching;
-		// important to keep the existing ds.__gdc cacheRef to prevent race condition causing a stale cacheRef
-		// to be used as ds.__gdc later
-		ds.__gdc.doneCaching = false // equivalent to having a ds.__pendingCacheVersion object present
-		ds.__pendingCacheVersion = version
-		ref.data_release_version = version
-		await getOpenProjects(ds, ref)
-	} catch (e) {
-		delete ds.__pendingCacheVersion
-		console.log('getOpenProjects() failed: ' + (e.message || e))
-		return
-	}
-
-	const begin = new Date()
-	try {
-		const size = 1000 // fetch 1000 ids at a time
-		const totalCases = await fetchIdsFromGdcApi(ds, 1, 0, ref)
-		if (!Number.isInteger(totalCases)) throw 'totalCases not integer'
-
-		console.log('GDC: Start to cache sample IDs of', totalCases, 'cases...')
-		for (let i = 0; i < Math.ceil(totalCases / size); i++) {
-			await fetchIdsFromGdcApi(ds, size, i * 1000, ref)
-		}
-
-		await getCasesWithGeneExpression(ds, ref)
-	} catch (e) {
-		console.log(e)
-		delete ds.__pendingCacheVersion
-		return
-	}
-
-	if (mayCancelStalePendingCache(ds, ref)) return
-	delete ds.__pendingCacheVersion
-	// swap to the newly completed cache reference
-	ds.__gdc = ref
-	ds.__gdc.doneCaching = true
-	console.log('GDC: Done caching sample IDs. Time:', Math.ceil((new Date() - begin) / 1000), 's')
-	console.log('\t', ds.__gdc.aliquot2submitter.cache.size, 'aliquot IDs to sample submitter id,')
-	console.log('\t', ds.__gdc.caseid2submitter.size, 'case uuid to submitter id,')
-	console.log('\t', ds.__gdc.map2caseid.cache.size, 'different ids to case uuid,')
-	console.log('\t', ds.__gdc.casesWithExpData.size, 'cases with gene expression data.')
-}
-
-// may cancel an unfinished caching for an older data_release_version,
-// if a new data_release_version is detected in cacheMappingOnNewRelease() runs
-function mayCancelStalePendingCache(ds, ref) {
-	//if (!ds.__pendingCacheVersion) return false
-	const next = ref.data_release_version
-	const current = ds.__pendingCacheVersion || ds.__gdc.data_release_version
-	if (!current) return false
-	if (next.major < current.major || next.minor < current.minor) {
-		console.log(`GDC: !!! cancel stale pending cache for `, next)
-		return true
-	}
-	return false
-}
-
-/*
-input:
-	ds:
-		gdc dataset object
-	size:int
-	from:int
-		null or integer
-		if null, aliquot_id must be given
-	aliquot_id:str
-
-output:
-	re.data.pagination.total
-
-aliquot-to-submitter mapping are automatically cached
-*/
-async function fetchIdsFromGdcApi(ds, size, from, ref, aliquot_id) {
-	if (mayCancelStalePendingCache(ds, ref)) return
-	const param = ['fields=submitter_id,samples.portions.analytes.aliquots.aliquot_id,samples.submitter_id']
-	if (aliquot_id) {
-		param.push(
-			'filters={"op":"and","content":[{"op":"=","content":{"field":"samples.portions.analytes.aliquots.aliquot_id","value":["' +
-				aliquot_id +
-				'"]}}]}'
-		)
-	} else {
-		if (!Number.isInteger(size) || !Number.isInteger(from)) throw 'size and from not integers'
-		param.push('size=' + size)
-		param.push('from=' + from)
-	}
-
-	const { host, headers } = ds.getHostHeaders()
-	const { body: re } = await cachedFetch(host.rest + '/cases?' + param.join('&'), { headers })
-	if (!Array.isArray(re?.data?.hits)) throw 're.data.hits[] not array'
-
-	//console.log(re.data.hits[0]) // uncomment to examine output
-
-	/*
-	re.data.hits = [
-	  {
-	  	// the "id" value seems to be case.case_id
-		// it is always here even if 'case.case_id' is not included in fields
-		"id": "c2829ab9-d5b2-5a82-a134-de9c591363de",
-		submitter_id: 'TCGA-LL-A6FQ', // case submitter id
-		"samples": [
-		  {
-			"submitter_id": "TARGET-50-PAJNID-01A",
-			"portions": [
-			  {
-				"analytes": [
-				  {
-					"aliquots": [
-					  {
-						"aliquot_id": "123bd4c3-6e36-4514-8d06-9f1f408cd1aa"
-					  }
-					]
-				  }
-				]
-			  },
-			  { ... more analytes ... }
-			]
-		  }
-		]
-	  }
-	 ]
-	*/
-	for (const h of re.data.hits) {
-		const case_id = h.id
-		if (!case_id) throw 'h.id (case uuid) missing'
-		ref.caseIds.add(case_id)
-
-		const case_submitter_id = h.submitter_id
-		if (!case_submitter_id) throw 'h.submitter_id missing'
-
-		ref.caseid2submitter.set(case_id, case_submitter_id)
-
-		/*
-		below shows different uuids mapping to same submitter id
-		this is the reason case submitter id must not be used to align data in oncomatrix, as it's not unique across Projects
-
-		if(ref.map2caseid.cache.has(case_submitter_id)) {
-			console.log(case_submitter_id, case_id, ref.map2caseid.cache.get(case_submitter_id))
-		}
-		*/
-
-		ref.map2caseid.cache.set(case_submitter_id, case_id)
-
-		if (!Array.isArray(h.samples)) continue //throw 'hit.samples[] not array'
-		for (const sample of h.samples) {
-			const sample_submitter_id = sample.submitter_id
-			if (!sample_submitter_id) throw 'sample.submitter_id missing'
-
-			ref.map2caseid.cache.set(sample_submitter_id, case_id)
-
-			if (!Array.isArray(sample.portions)) continue // throw 'sample.portions[] not array'
-			for (const portion of sample.portions) {
-				if (!Array.isArray(portion.analytes)) continue //throw 'portion.analytes not array'
-				for (const analyte of portion.analytes) {
-					if (!Array.isArray(analyte.aliquots)) continue //throw 'analyte.aliquots not array'
-					for (const aliquot of analyte.aliquots) {
-						const aliquot_id = aliquot.aliquot_id
-						if (!aliquot_id) throw 'aliquot.aliquot_id missing'
-						ref.aliquot2submitter.cache.set(aliquot_id, sample_submitter_id)
-						ref.map2caseid.cache.set(aliquot_id, case_id)
-					}
-				}
-			}
-		}
-	}
-	return re.data?.pagination?.total
-}
-
-async function getCasesWithGeneExpression(ds, ref) {
-	if (mayCancelStalePendingCache(ds, ref)) return
-	const { host, headers } = ds.getHostHeaders()
-	const url = path.join(host.geneExp, 'gene_expression/availability')
-
-	try {
-		const idLst = [...ref.caseIds]
-		const { body: re } = await cachedFetch(url, {
-			method: 'post',
-			headers,
-			body: { case_ids: idLst, gene_ids: ['ENSG00000141510'] }
-		})
-		// {"cases":{"details":[{"case_id":"4abbd258-0f0c-4428-901d-625d47ad363a","has_gene_expression_values":true}],"with_gene_expression_count":1,"without_gene_expression_count":0},"genes":null}
-		if (!Array.isArray(re.cases?.details)) throw 're.cases.details[] not array'
-		for (const c of re.cases.details) {
-			if (c.has_gene_expression_values) ref.casesWithExpData.add(c.case_id)
-		}
-
-		delete ref.caseIds
-	} catch (e) {
-		console.log(e)
-		console.log("You don't have access to /gene_expression/availability/, you cannot run GDC hierCluster")
-		delete ds.__pendingCacheVersion
-		// while gene exp api are only on uat-prod, do not throw here so a pp instance with IP not whitelisted on uat-prod will not be broken
-		// when api is releated to public prod, must throw here and abort
 	}
 }
