+++ server/src/mds3.gdc.js
@@ -2,7 +2,6 @@ import * as common from '#shared/common.js'
 import { compute_bins } from '#shared/termdb.bins.js'
 import ky from 'ky'
 import nodeFetch from 'node-fetch'
-import path from 'path'
 import { combineSamplesById } from './mds3.variant2samples.js'
 import { filter2GDCfilter } from './mds3.gdc.filter.js'
 import { write_tmpfile } from './utils.js'
