+++ server/src/mds3.init.js
@@ -6,7 +6,7 @@ import { scaleLinear } from 'd3-scale'
 import { run_rust } from '@sjcrh/proteinpaint-rust'
 import { createCanvas } from 'canvas'
 import * as gdc from './mds3.gdc.js'
-import { initGDCdictionary } from './termdb.gdc.js'
+import { initGDCdictionary } from './initGdc.js'
 import { validate_variant2samples } from './mds3.variant2samples.js'
 import { ssmIdFieldsSeparator } from '#shared/mds3tk.js'
 import * as utils from './utils.js'
